joshu/requirements.txt
amqp==5.0.2
APScheduler==3.6.3
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.2
celery-progress==0.0.14
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
DAWG-Python==0.7.2
decorator==4.4.2
Django==3.1.3
django-bootstrap4==2.3.1
django-celery-results==2.0.0
django-common-helpers==0.9.2
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
drf-yasg==1.20.0
emoji==0.6.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
numpy==1.19.4
packaging==20.7
prompt-toolkit==3.0.8
pycparser==2.20
PyJWT==1.7.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
pyTelegramBotAPI==3.7.3
python-telegram-bot==13.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.7.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
telebot==0.0.3
telebot-calendar==1.1
tornado==6.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
